Resolve some TODOs [DPP-876]
changelog_begin
changelog_end
