+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/StandaloneIndexService.scala
@@ -58,7 +58,6 @@ object StandaloneIndexService {
     }
 
     def loadDamlPackages(): InMemoryPackageStore = {
-      // TODO is it sensible to have all the initial packages to be known since the epoch?
       config.archiveFiles
         .foldLeft[Either[(String, File), InMemoryPackageStore]](Right(InMemoryPackageStore.empty)) {
           case (storeE, f) =>
