--- ledger/participant-integration-api/src/main/scala/platform/store/appendonlydao/events/EventsTableQueries.scala
@@ -1,11 +1,0 @@
-// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.platform.store.appendonlydao.events
-
-private[events] object EventsTableQueries {
-  def format(ps: Set[Party]): String =
-    ps.view.map(p => s"'$p'").mkString(",")
-  def format(ps: List[Party]): String =
-    ps.map(p => s"'$p'").mkString(",")
-}
