+++ ledger/participant-integration-api/src/main/scala/platform/store/appendonlydao/events/package.scala
@@ -3,9 +3,6 @@
 
 package com.daml.platform.store.appendonlydao
 
-import akka.stream.scaladsl.Source
-import com.daml.lf.transaction.Node.KeyWithMaintainers
-
 // TODO append-only: revisit visibility, and necessity during cleanup
 /** Type aliases used throughout the package
   */
@@ -58,43 +55,4 @@ package object events {
 
   import com.daml.lf.crypto
   type Hash = crypto.Hash
-
-  /** Groups together items of type [[A]] that share an attribute [[K]] over a
-    * contiguous stretch of the input [[Source]]. Well suited to perform group-by
-    * operations of streams where [[K]] attributes are either sorted or at least
-    * show up in blocks.
-    *
-    * Implementation detail: this method _must_ use concatSubstreams instead of
-    * mergeSubstreams to prevent the substreams to be processed in parallel,
-    * potentially causing the outputs to be delivered in a different order.
-    *
-    * Docs: https://doc.akka.io/docs/akka/2.6.10/stream/stream-substream.html#groupby
-    */
-  def groupContiguous[A, K, Mat](
-      source: Source[A, Mat]
-  )(by: A => K): Source[Vector[A], Mat] =
-    source
-      .statefulMapConcat(() => {
-        var previousSegmentKey: K = null.asInstanceOf[K]
-        entry => {
-          val keyForEntry = by(entry)
-          val entryWithSplit = entry -> (keyForEntry != previousSegmentKey)
-          previousSegmentKey = keyForEntry
-          List(entryWithSplit)
-        }
-      })
-      .splitWhen(_._2)
-      .map(_._1)
-      .fold(Vector.empty[A])(_ :+ _)
-      .concatSubstreams
-
-  def convert(template: Identifier, key: lftx.Node.KeyWithMaintainers): Key =
-    Key.assertBuild(template, key.key)
-
-  def convertLfValueKey(
-      template: Identifier,
-      key: KeyWithMaintainers,
-  ) =
-    Key.assertBuild(template, key.key)
-
 }
